Replace git-latest and hg-latest with latest rule
Combines two similar rules into a more generic one that can be used
for similar purposed.
