+++ examples/playbooks/skiptasks.yml
@@ -2,10 +2,10 @@
 - hosts: all
 
   tasks:
-    - name: Test git-latest
+    - name: Test latest[git]
       action: ansible.builtin.git
 
-    - name: Test hg-latest
+    - name: Test latest[hg]
       action: ansible.builtin.hg
 
     - name: Test command-instead-of-module
@@ -15,12 +15,12 @@
     - name: Test deprecated-command-syntax
       ansible.builtin.command: creates=B chmod 644 A
 
-    - name: Test git-latest (skip)
+    - name: Test latest[git] (skip)
       action: ansible.builtin.git
       tags:
         - skip_ansible_lint
 
-    - name: Test hg-latest (skip)
+    - name: Test latest[hg] (skip)
       action: ansible.builtin.hg
       tags:
         - skip_ansible_lint
@@ -35,30 +35,30 @@
       tags:
         - skip_ansible_lint
 
-    - name: Test git-latest (don't warn)
+    - name: Test latest[git] (don't warn)
       ansible.builtin.command: git log
       args:
         warn: false
       changed_when: false
 
-    - name: Test hg-latest (don't warn)
+    - name: Test latest[hg] (don't warn)
       ansible.builtin.command: chmod 644 A
       args:
         warn: false
         creates: B
 
-    - name: Test hg-latest (warn)
+    - name: Test latest[hg] (warn)
       ansible.builtin.command: chmod 644 A
       args:
         warn: true
         creates: B
 
-    - name: Test git-latest (don't warn single line)
+    - name: Test latest[git] (don't warn single line)
       ansible.builtin.command: warn=False chdir=/tmp/blah git log
       changed_when: false
 
-    - name: Test hg-latest (don't warn single line)
+    - name: Test latest[hg] (don't warn single line)
       ansible.builtin.command: warn=no creates=B chmod 644 A
 
-    - name: Test hg-latest (warn single line)
+    - name: Test latest[hg] (warn single line)
       ansible.builtin.command: warn=yes creates=B chmod 644 A
