+++ src/ansiblelint/schemas/ansible-lint-config.json
@@ -94,8 +94,7 @@
           "deprecated-module",
           "empty-string-compare",
           "fqcn-builtins",
-          "git-latest",
-          "hg-latest",
+          "latest",
           "ignore-errors",
           "inline-env-var",
           "internal-error",
