+++ test/test_skip_playbook_items.py
@@ -8,12 +8,12 @@ PLAYBOOK_PRE_TASKS = """\
 - name: Fixture
   hosts: all
   tasks:
-    - name: Bad git 1  # noqa git-latest
+    - name: Bad git 1  # noqa latest[git]
       action: ansible.builtin.git a=b c=d
     - name: Bad git 2
       action: ansible.builtin.git a=b c=d
   pre_tasks:
-    - name: Bad git 3  # noqa git-latest
+    - name: Bad git 3  # noqa latest[git]
       action: ansible.builtin.git a=b c=d
     - name: Bad git 4
       action: ansible.builtin.git a=b c=d
@@ -24,12 +24,12 @@ PLAYBOOK_POST_TASKS = """\
 - name: Fixture
   hosts: all
   tasks:
-    - name: Bad git 1  # noqa git-latest
+    - name: Bad git 1  # noqa latest[git]
       action: ansible.builtin.git a=b c=d
     - name: Bad git 2
       action: ansible.builtin.git a=b c=d
   post_tasks:
-    - name: Bad git 3  # noqa git-latest
+    - name: Bad git 3  # noqa latest[git]
       action: ansible.builtin.git a=b c=d
     - name: Bad git 4
       action: ansible.builtin.git a=b c=d
@@ -40,12 +40,12 @@ PLAYBOOK_HANDLERS = """\
 - name: Fixture
   hosts: all
   tasks:
-    - name: Bad git 1  # noqa git-latest
+    - name: Bad git 1  # noqa latest[git]
       action: ansible.builtin.git a=b c=d
     - name: Bad git 2
       action: ansible.builtin.git a=b c=d
   handlers:
-    - name: Bad git 3  # noqa git-latest
+    - name: Bad git 3  # noqa latest[git]
       action: ansible.builtin.git a=b c=d
     - name: Bad git 4
       action: ansible.builtin.git a=b c=d
@@ -56,7 +56,7 @@ PLAYBOOK_TWO_PLAYS = """\
 - name: Fixture
   hosts: all
   tasks:
-    - name: Bad git 1  # noqa git-latest
+    - name: Bad git 1  # noqa latest[git]
       action: ansible.builtin.git a=b c=d
     - name: Bad git 2
       action: ansible.builtin.git a=b c=d
@@ -64,7 +64,7 @@ PLAYBOOK_TWO_PLAYS = """\
 - name: Fixture 2
   hosts: all
   tasks:
-    - name: Bad git 3  # noqa git-latest
+    - name: Bad git 3  # noqa latest[git]
       action: ansible.builtin.git a=b c=d
     - name: Bad git 4
       action: ansible.builtin.git a=b c=d
@@ -75,23 +75,23 @@ PLAYBOOK_WITH_BLOCK = """\
 - name: Fixture
   hosts: all
   tasks:
-    - name: Bad git 1  # noqa git-latest
+    - name: Bad git 1  # noqa latest[git]
       action: ansible.builtin.git a=b c=d
     - name: Bad git 2
       action: ansible.builtin.git a=b c=d
     - name: Block with rescue and always section
       block:
-        - name: Bad git 3  # noqa git-latest
+        - name: Bad git 3  # noqa latest[git]
           action: ansible.builtin.git a=b c=d
         - name: Bad git 4
           action: ansible.builtin.git a=b c=d
       rescue:
-        - name: Bad git 5  # noqa git-latest
+        - name: Bad git 5  # noqa latest[git]
           action: ansible.builtin.git a=b c=d
         - name: Bad git 6
           action: ansible.builtin.git a=b c=d
       always:
-        - name: Bad git 7  # noqa git-latest
+        - name: Bad git 7  # noqa latest[git]
           action: ansible.builtin.git a=b c=d
         - name: Bad git 8
           action: ansible.builtin.git a=b c=d
